requirements.txt
# Core dependencies
streamlit>=1.28.0
datasets>=2.14.0,<3.0.0
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning & NLP
sentence-transformers>=2.2.0,<2.3.0
google-generativeai>=0.3.0
transformers>=4.21.0,<4.22.0
torch>=2.0.0
huggingface-hub>=0.15.0,<0.16.0

# Graph Database
neo4j>=5.0.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0

# Development and testing
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0

# Optional: For advanced evaluation metrics
rouge-score>=0.1.2
bert-score>=0.3.13
